Rename named_skill_tactics to assigned_skill_tactics
