+++ src/software/ai/hl/stp/skill/dribble/dribble_skill_test.cpp
@@ -2,7 +2,7 @@
 
 #include <utility>
 
-#include "software/ai/hl/stp/tactic/assigned_skill/named_skill_tactics.h"
+#include "software/ai/hl/stp/tactic/assigned_skill/assigned_skill_tactics.h"
 #include "software/geom/algorithms/contains.h"
 #include "software/simulated_tests/non_terminating_validation_functions/robot_not_excessively_dribbling_validation.h"
 #include "software/simulated_tests/simulated_er_force_sim_play_test_fixture.h"
