+++ src/software/ai/hl/stp/skill/pivot_kick/BUILD
@@ -36,7 +36,7 @@ cc_test(
     deps = [
         ":pivot_kick_skill",
         "//shared/test_util:tbots_gtest_main",
-        "//software/ai/hl/stp/tactic/assigned_skill:named_skill_tactics",
+        "//software/ai/hl/stp/tactic/assigned_skill:assigned_skill_tactics",
         "//software/simulated_tests:simulated_er_force_sim_play_test_fixture",
         "//software/simulated_tests/terminating_validation_functions",
         "//software/simulated_tests/validation:validation_function",
