+++ src/software/ai/hl/stp/tactic/BUILD
@@ -16,7 +16,7 @@ cc_library(
     deps = [
         ":tactic",
         "//software/ai/hl/stp/tactic/assigned_skill:assigned_skill_tactic",
-        "//software/ai/hl/stp/tactic/assigned_skill:named_skill_tactics",
+        "//software/ai/hl/stp/tactic/assigned_skill:assigned_skill_tactics",
         "//software/ai/hl/stp/tactic/attacker:attacker_tactic",
         "//software/ai/hl/stp/tactic/chip:chip_tactic",
         "//software/ai/hl/stp/tactic/crease_defender:crease_defender_tactic",
