+++ src/software/ai/hl/stp/tactic/all_tactics.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "software/ai/hl/stp/tactic/assigned_skill/assigned_skill_tactic.hpp"
-#include "software/ai/hl/stp/tactic/assigned_skill/named_skill_tactics.h"
+#include "software/ai/hl/stp/tactic/assigned_skill/assigned_skill_tactics.h"
 #include "software/ai/hl/stp/tactic/attacker/attacker_tactic.h"
 #include "software/ai/hl/stp/tactic/chip/chip_tactic.h"
 #include "software/ai/hl/stp/tactic/crease_defender/crease_defender_tactic.h"
