+++ src/software/ai/hl/stp/tactic/assigned_skill/BUILD
@@ -11,8 +11,8 @@ cc_library(
 )
 
 cc_library(
-    name = "named_skill_tactics",
-    hdrs = ["named_skill_tactics.h"],
+    name = "assigned_skill_tactics",
+    hdrs = ["assigned_skill_tactics.h"],
     deps = [
         ":assigned_skill_tactic",
         "//software/ai/hl/stp/skill/pivot_kick:pivot_kick_skill",
